+++ src/app/App.tsx
@@ -1,22 +1,10 @@
 import React from "react";
 import MainPage from "../pages/MainPage";
-import "./App.css";
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
         <MainPage />
       </header>
     </div>
