--- src/shared/Button/Button.css
@@ -1,30 +1,0 @@
-.custom-button {
-  border: none;
-  border-radius: 1px;
-  padding: 10px 16px;
-  font-size: 14px;
-  font-weight: 500;
-  cursor: pointer;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-  transition: all 0.2s ease-in-out;
-
-  box-sizing: border-box;
-}
-
-.custom-button:hover {
-  opacity: 0.9;
-  transform: translateY(-1px);
-}
-
-.custom-button:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-  transform: none;
-}
-
-.button-icon {
-  display: flex;
-}
