+++ src/shared/button/IconButton.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "./Button/Button";
+import Button from "./Button";
 
 interface IconButtonProps
   extends Omit<
