+++ src/shared/css/main.css
@@ -1,0 +1,7 @@
+:root {
+    --main-color-blue: #2A3FA7;
+    --main-color-orange: #FF9500;
+
+    --main-text-black: #1E1E22;
+    --main-warn-color: #E81414;
+}
